Add Page model and boot-script that shows the issue
There seems to be an issue with the validatesUniquenessOf method.

The method only gets triggered when an individual item gets created.
When an array of items gets created, the rule will be by-passed.

In this example there is a Page model with one required property 'title'

There is a validatesUniquenessOf rule that should make sure the title
field is unique.

The bootscript `add-pages.js` shows that this rule gets ignored when
an array of items is created. It does work when creating individual ones.